+++ WorkManagerDal/Services/IRolesService.cs
@@ -8,9 +8,9 @@ namespace WorkManagerDal.Services
         Task<bool> IsRoleExistAsync(string name);
         Task<Role> CreateRoleAsync(Role role);
         Task<Role> UpdateRoleAsync(Role role);
-        Task<Role> GetRoleWithPermissionsAsync(int id);
-        Task DeleteRoleToPemissionAsync(int roleId, string permissionName);
-        Task AddRoleToPermissionAsync(int roleId, string permissionName);
+        Task<Role> GetRoleWithPermissionsAsync(long id);
+        Task DeleteRoleToPemissionAsync(long roleId, string permissionName);
+        Task AddRoleToPermissionAsync(long roleId, string permissionName);
         Task DeleteRoleAsync(Role role);
     }
 }
